+++ service/grails-app/controllers/org/olf/licenses/ApplicationController.groovy
@@ -1,8 +1,9 @@
-package org.olf
+package org.olf.licenses
+
+import org.olf.licenses.DataloadService
 
 import grails.core.GrailsApplication
 import grails.plugins.*
-import org.olf.license.DataloadService
 import grails.gorm.multitenancy.CurrentTenant
 import groovy.util.logging.Slf4j
