--- service/grails-app/controllers/org/olf/licenses/TestController.groovy
@@ -1,11 +1,0 @@
-package org.olf.licenses
-
-
-import grails.rest.*
-import grails.converters.*
-
-class TestController {
-	static responseFormats = ['json', 'xml']
-	
-    def index() { }
-}
