+++ service/grails-app/views/license/_license.gson
@@ -1,5 +1,6 @@
-import groovy.transform.*
 import org.olf.licenses.License
+
+import groovy.transform.*
 
 @Field License license
 json g.render(license, [expand: ['licenseProps','tags']])
