--- service/src/test/groovy/org/olf/licenses/TestControllerSpec.groovy
@@ -1,19 +1,0 @@
-package org.olf.licenses
-
-import geb.spock.GebSpec
-import grails.testing.web.controllers.ControllerUnitTest
-import spock.lang.Specification
-
-class TestControllerSpec extends GebSpec implements ControllerUnitTest<TestController> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-        true == false
-    }
-}
